requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
binance==0.3
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
discord==1.0.1
discord.py==1.7.2
idna==2.10
matlab==0.1
multidict==5.1.0
numpy==1.20.2
panda==0.3.1
pandas==1.2.4
py==1.10.0
python-binance==1.0.10
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scipy==1.6.3
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
urlopen==1.0.0
websockets==9.0.2
yarl==1.6.3
